# crates/hs-admin/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: hs-admin API
  summary: The native administration and moderation API for the homeserver.
  version: 0.1.0
  description: 'The only API the management interface (track 16) talks to. See docs/rfcs/0004-admin-api.md for the full design rationale: resource modeling, naming, pagination, filtering, errors, idempotency, scopes, versioning and the event stream. This document is the contract; the hs-admin router''s contract tests assert the router and this document agree on paths and methods (RFC 0004 section 1, decision D15.16).'
  contact:
    name: 'Track 15: Admin API and modules'
    url: https://github.com/example/matrix-reimplement/blob/main/docs/workstreams/15-admin-api-and-modules.md
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  x-hs-contract: 1.0-draft
  x-hs-changelog:
  - version: 0.1.0
    date: '2026-09-18'
    notes: Initial draft covering every resource in RFC 0004 section 4. Not yet frozen; freezes at week 8 as contract 1.0.
servers:
- url: /api/v1
  description: Same origin as /_matrix on the client listener.
...
security:
- OAuth2: []
paths:
...
  /rooms/{room_id}/aliases:
    get:
      summary: List a room's aliases
      operationId: rooms.aliases.list
      tags:
      - Rooms
      security:
      - OAuth2:
        - moderation:read
      parameters:
      - $ref: '#/components/parameters/RoomId'
      responses:
        '200':
          description: Aliases.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoomAlias'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/InsufficientScope'
        '429':
          $ref: '#/components/responses/RateLimited'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Internal'
        '503':
          $ref: '#/components/responses/Unavailable'
    post:
      summary: Add an alias
      operationId: rooms.aliases.add
      tags:
      - Rooms
      security:
      - OAuth2:
        - moderation:write
      parameters:
      - $ref: '#/components/parameters/RoomId'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - alias
              properties:
                alias:
                  type: string
      responses:
        '201':
          description: Added.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomAlias'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/InsufficientScope'
        '429':
          $ref: '#/components/responses/RateLimited'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/IdempotencyMismatch'
        '500':
          $ref: '#/components/responses/Internal'
        '503':
          $ref: '#/components/responses/Unavailable'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
  /rooms/{room_id}/aliases/{alias}:
    delete:
      summary: Remove an alias
      operationId: rooms.aliases.remove
      tags:
      - Rooms
      security:
      - OAuth2:
        - moderation:write
      parameters:
      - $ref: '#/components/parameters/RoomId'
      - name: alias
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Deleted.
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/InsufficientScope'
        '429':
          $ref: '#/components/responses/RateLimited'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Internal'
        '503':
          $ref: '#/components/responses/Unavailable'
...
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      description: Tokens issued by 07's OAuth 2.0 issuer (authorization code with PKCE for operators, client credentials for automation). A legacy Matrix access token of a user with the server-administrator flag is also accepted and is treated as admin:read plus admin:write.
      flows:
        authorizationCode:
          authorizationUrl: https://example.org/_matrix/client/oauth2/authorize
          tokenUrl: https://example.org/_matrix/client/oauth2/token
          scopes:
            admin:read: Read every resource, including the audit log, configuration (secrets redacted), cluster and migration status. Satisfies every other :read scope.
            admin:write: Everything. Implies every other scope.
            bridges:read: Read appservices, bridge types, their health and backlog, and appservice tasks and events.
            bridges:write: bridges:read plus create, update, delete, pause, resume, replay, rotate tokens, export registrations, render bridge types.
            moderation:read: Read users, rooms, reports, media, their sub-resources, moderation tasks and events. Sessions' IP addresses and account data are redacted. Not message content (a room's timeline and its events), which needs admin:read (decision 0013).
            moderation:write: moderation:read plus suspend, lock, shadow-ban, deactivate, redact, reset password, logout, room block/delete/purge, quarantine and delete media, resolve reports, server notices.
        clientCredentials:
          tokenUrl: https://example.org/_matrix/client/oauth2/token
          scopes:
            admin:read: Read every resource, including the audit log, configuration (secrets redacted), cluster and migration status. Satisfies every other :read scope.
            admin:write: Everything. Implies every other scope.
            bridges:read: Read appservices, bridge types, their health and backlog, and appservice tasks and events.
            bridges:write: bridges:read plus create, update, delete, pause, resume, replay, rotate tokens, export registrations, render bridge types.
            moderation:read: Read users, rooms, reports, media, their sub-resources, moderation tasks and events. Sessions' IP addresses and account data are redacted. Not message content (a room's timeline and its events), which needs admin:read (decision 0013).
            moderation:write: moderation:read plus suspend, lock, shadow-ban, deactivate, redact, reset password, logout, room block/delete/purge, quarantine and delete media, resolve reports, server notices.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: 1-255 visible ASCII characters. Replays return the stored response for 24 hours.
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 255
...
    RoomId:
      name: room_id
      in: path
      description: A Matrix room id, percent-encoded.
      required: true
      schema:
        type: string
      example: '%21abc%3Aexample.org'
...
  responses:
    ValidationFailed:
      description: Malformed JSON, schema violation, or bad parameter (urn:hs:problem:validation-failed).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Unauthenticated:
      description: Missing, expired, revoked or malformed token (urn:hs:problem:unauthenticated).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        WWW-Authenticate:
          schema:
            type: string
          description: Bearer realm="hs-admin", error="invalid_token"
    InsufficientScope:
      description: The token lacks the operation's required scope (urn:hs:problem:insufficient-scope).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        WWW-Authenticate:
          schema:
            type: string
          description: Bearer error="insufficient_scope", scope="..."
...
    NotFound:
      description: Unknown path or resource (urn:hs:problem:not-found).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Duplicate identifier, namespace conflict, or state conflict (urn:hs:problem:conflict).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    PayloadTooLarge:
      description: Request body exceeds the configured size limit (urn:hs:problem:payload-too-large).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    UnsupportedMediaType:
      description: A body was sent with a Content-Type other than application/json (urn:hs:problem:unsupported-media-type).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    IdempotencyMismatch:
      description: The Idempotency-Key was reused with a different request (urn:hs:problem:idempotency-key-payload-mismatch).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    RateLimited:
      description: Too many requests (urn:hs:problem:rate-limited).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds.
    Internal:
      description: Unexpected server error (urn:hs:problem:internal). Never includes internals.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Unavailable:
      description: Owning replica unreachable, store unavailable, or read-only during cutover (urn:hs:problem:unavailable).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details, used for every /api/v1 error.
      required:
      - type
      - title
      - status
      properties:
        type:
          type: string
          format: uri
          description: urn:hs:problem:<slug> from the closed catalog in RFC 0004 section 3.5.
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        request_id:
          type: string
          description: ULID; matches X-Request-Id and tracing spans.
        errcode:
          type: string
          description: The Matrix error code, present whenever a Matrix client could observe the same failure.
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
        retry_after_ms:
          type: integer
        required_scope:
          type: string
    ValidationError:
      type: object
      required:
      - pointer
      - detail
      properties:
        pointer:
          type: string
          description: JSON Pointer into the body, or param:<name> for a query parameter.
        detail:
          type: string
...
    RoomAlias:
      type: object
      required:
      - alias
      - creator
      - canonical
      properties:
        alias:
          type: string
        created_at:
          type:
          - string
          - 'null'
          format: date-time
          description: Not recorded for aliases made before it was; null then.
        creator:
          type:
          - string
          - 'null'
          description: Who made the alias, when that is known.
        canonical:
          type: boolean
          description: Whether the room's m.room.canonical_alias names it.
...